Ticket [https://trac.osgeo.org/fdo/ticket/625]: More fixes to be able to build FDO against VS2010  - Fix a warning
git-svn-id: http://svn.osgeo.org/fdo/trunk@5978 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Utilities/Common/Inc/FdoCommonHash.h
@@ -183,7 +183,7 @@ private:
             lists[i] = NULL;
         HashList** tmp = m_lists;
         m_lists = lists;
-        for (int i = 0; i < oldCap; i++)
+        for (size_t i = 0; i < oldCap; i++)
         {
             HashList* oldLst = tmp[i];
             if (oldLst != NULL)
